+++ crates/tests/src/ty_escape.rs
@@ -5,7 +5,7 @@ fn smoke() {
   check(
     r#"
 val _ = let datatype bad = guh in guh end
-(**     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ type escapes its scope: bad *)
+(**     ^^^ type escapes its scope: bad *)
 "#,
   );
 }
@@ -15,7 +15,7 @@ fn seq() {
   check(
     r#"
 val _ = let datatype foo = bar val quz = bar in (bar; 3 + 3; quz) end
-(**     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ type escapes its scope: foo *)
+(**     ^^^ type escapes its scope: foo *)
 "#,
   );
 }
@@ -34,7 +34,7 @@ fn branch() {
   check(
     r#"
 val _ = let datatype d = D in if 3 < 4 then [] else [(3, D, "hi")] end
-(**     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ type escapes its scope: d *)
+(**     ^^^ type escapes its scope: d *)
 "#,
   );
 }
@@ -45,7 +45,7 @@ fn shadow() {
     r#"
 datatype t = One
 val _ = let datatype t = Two in Two end
-(**     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ type escapes its scope: t *)
+(**     ^^^ type escapes its scope: t *)
 "#,
   );
 }